// src/app/modelo/enum/accion-favor-enum.ts
export enum AccionFavorEnum {
    cancelaSolicitante = "cancelaSolicitante",
    aceptaSolicitante = "aceptaSolicitante",
    rechazaSolicitante = "rechazaSolicitante",
    finalizaSolicitante = "finalizaSolicitante",
    calificaSolicitante = "calificaSolicitante",
    aceptaAyudante = "aceptaAyudante",
    cancelaAyudante = "cancelaAyudante",
    finalizaAyudante = "finalizaAyudante",
    calificaAyudante = "calificaAyudante",
}